python/requirements.txt
pyyaml==6.0.2
requests==2.32.3
edgartools==3.2.1
anyio==4.6.2.post1
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-ai-ml==1.22.2
python-dotenv==1.0.1
yfinance==0.2.40
semantic-kernel==1.17.1
azure-ai-evaluation
flake8-docstrings==1.7.0
flake8==7.1.1
pep8-naming==0.14.1
pytest==8.3.3
pytest-cov==5.0.0
## otel collector
opentelemetry-instrumentation-openai==0.33.12
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-distro==0.49b2
opentelemetry-exporter-otlp==1.28.2
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.8.2
## set this only when you want to directly instrument using
## application insights without collector
azure-monitor-opentelemetry==1.6.4
